+++ inst/doc/create_project.R
@@ -29,48 +29,58 @@ readr::read_file(system.file("project_templates/default_readme.md", package = 's
 
 ## -----------------------------------------------------------------------------
 my_template <-
-  quote(list(
+  list(
     gitignore = list(),
     readme = list(),
     rproj = list()
-  ))
+  )
 
 ## ---- eval=FALSE--------------------------------------------------------------
-#  readme <- list(
-#    template_filename = system.file("project_templates/default_readme.md", package = "starter"),
-#    filename = "README.md",
-#    glue = TRUE
-#  )
+#  readme <-
+#    list(
+#      template_filename = system.file("project_templates/default_readme.md", package = "starter"),
+#      filename = "README.md",
+#      glue = TRUE
+#    )
 
 ## ---- eval=FALSE--------------------------------------------------------------
-#  rproj <- list(
-#    template_filename = system.file("project_templates/default_rproj.Rproj", package = "starter"),
-#    filename = glue::glue("{folder_name}.Rproj"),
-#    glue = FALSE
-#  )
+#  rproj <-
+#    rlang::expr(list(
+#      template_filename = system.file("project_templates/default_rproj.Rproj", package = "starter"),
+#      filename = glue::glue("{folder_name}.Rproj"),
+#      glue = FALSE
+#    ))
 
 ## -----------------------------------------------------------------------------
-my_template <- quote(list(
-  readme = list(
-    template_filename = system.file("project_templates/default_readme.md", package = "starter"), 
-    filename = "README.md",
-    glue = TRUE
-  ),
-  rproj = list(
-    template_filename = system.file("project_templates/default_rproj.Rproj", package = "starter"), 
-    filename = glue::glue("{folder_name}.Rproj"),
-    glue = FALSE
-  ),
-  
-gitignore = list(
-  template_filename = system.file("project_templates/default_gitignore.txt", package = "starter"), 
-  filename = "gitignore.txt", 
-  glue = TRUE
-)
-))
+my_template <- 
+  list(
+    readme = rlang::expr(list(
+      template_filename = system.file("project_templates/default_readme.md", package = "starter"), 
+      filename = "README.md",
+      glue = TRUE
+    )),
+    rproj = rlang::expr(list(
+      template_filename = system.file("project_templates/default_rproj.Rproj", package = "starter"), 
+      filename = glue::glue("{folder_name}.Rproj"),
+      glue = FALSE
+    )),
+    gitignore = rlang::expr(list(
+      template_filename = system.file("project_templates/default_gitignore.txt", package = "starter"), 
+      filename = "gitignore.txt", 
+      glue = TRUE
+    ))
+  )
+
+## ---- eval = FALSE------------------------------------------------------------
+#  analysis <-
+#    list(
+#      template_filename = "C:/project_templates/template_analysis.Rmd",
+#      filename = "scripts/analysis.Rmd",
+#      glue = FALSE
+#    )
 
 ## -----------------------------------------------------------------------------
-attr(my_template, "label") <- "My Very 1st Project Template"
+attr(my_template, "label") <- "My 1st Project Template"
 
 ## ---- include=FALSE, echo=FALSE-----------------------------------------------
 unlink(project_path, recursive = TRUE)